Clicking on pass in error state re-attempts getting the password
When user cancels the passphrase request or enters wrong passphrase
three times, the password entry in the view is stuck in error state
for ever with no way to retrieve the password again until the applet
is restarted.

This change allows the password entry to be clicked even when it is
in the error state and it will try to obtain the password again.

BUG: 435825
FIXED-IN: 1.2.1