Fix entire clipboard history being always wiped by Plasma Pass
Modern Klipper supports x-kde-passwordManagerHint and won't store
items marked with this hint in the clipboard history. Back when
I was writing Plasma Pass this feature wasn't merged yet. Plasma
Pass was ready for it but it also had to support the current version
of Klipper at that time, so it always go and try to remove the password
from Klipper's history.

When new Klipper with support for x-kde-passwordManagerHint was released
Plasma Pass would still try to remove the password from Klipper's
clipboard history but would always fail (since Klipper doesn't store
passwords in history anymore). This would trigger a safety fallback
inside Plasma Pass that just wipes entire Klipper history, just to
be sure.

With this fix, Plasma Pass will only attempt to remove the password
from Klipper's history if it detects an old Klipper that doesn't support
the hint yet.

BUG: 405135
FIXED-IN: 1.2.1